proposing config not in profile
