+++ conans/test/functional/configuration/test_conf_file.py
@@ -1,0 +1,30 @@
+import textwrap
+
+import pytest
+
+from conans.test.utils.tools import TestClient
+
+
+@pytest.fixture
+def client():
+    client = TestClient()
+    conanfile = textwrap.dedent("""
+        from conans import ConanFile
+
+        class Pkg(ConanFile):
+            def generate(self):
+                for k, conf in self.conf.items():
+                    for name, value in conf.items():
+                        self.output.info("{}${}${}".format(k, name, value))
+        """)
+    client.save({"conanfile.py": conanfile})
+    return client
+
+
+def test_msbuild_config(client):
+    conf = textwrap.dedent("""\
+        tools.microsoft.MSBuild:verbosity=Minimal
+        """)
+    client.save({"conf_file": conf})
+    client.run("create . pkg/0.1@ --config-file=conf_file")
+    assert "/verbosity:Minimal" in client.out
